Moving show.py to shows.py to keep consistency*
